+++ leaderboardFeature/src/main/java/com/panhuk/leaderboardfeature/utils/Image.kt
@@ -1,20 +1,21 @@
-package com.panhuk.leaderboardfeature
+package com.panhuk.leaderboardfeature.utils
 
+import com.panhuk.leaderboardfeature.R.drawable
 import kotlin.random.Random
 
 fun getDrawable(): Int {
   return when (Random.nextInt(11)) {
-    0 -> R.drawable.add1
-    1 -> R.drawable.add2
-    2 -> R.drawable.android1
-    3 -> R.drawable.android2
-    4 -> R.drawable.banana
-    5 -> R.drawable.cash
-    6 -> R.drawable.crown
-    7 -> R.drawable.favorite
-    8 -> R.drawable.footprint
-    9 -> R.drawable.goal
-    10 -> R.drawable.handicrafts
-    else -> R.drawable.android2
+    0 -> drawable.add1
+    1 -> drawable.add2
+    2 -> drawable.android1
+    3 -> drawable.android2
+    4 -> drawable.banana
+    5 -> drawable.cash
+    6 -> drawable.crown
+    7 -> drawable.favorite
+    8 -> drawable.footprint
+    9 -> drawable.goal
+    10 -> drawable.handicrafts
+    else -> drawable.android2
   }
 }
